chore(refactor): refactor the routes
This will allow for selective CORs setting. Public routes like `/proxy`
will have the CORS allow origin header match the client header but for
auth routes, and others that will come in future releases, this will not
be the case.
